Issue-3499: Move service module to compose/core package
Partially implements #3499
