+++ compose/cli/main.py
@@ -23,11 +23,6 @@ from ..config.serialize import serialize_config
 from ..core.project import NoSuchService
 from ..core.project import OneOffFilter
 from ..core.project import ProjectError
-from ..service import BuildAction
-from ..service import BuildError
-from ..service import ConvergenceStrategy
-from ..service import ImageType
-from ..service import NeedsBuildError
 from .command import get_config_path_from_options
 from .command import project_from_options
 from .docopt_command import DocoptDispatcher
@@ -42,6 +37,11 @@ from .utils import get_version_info
 from .utils import yesno
 from compose.common import signals
 from compose.common.progress_stream import StreamOutputError
+from compose.core.service import BuildAction
+from compose.core.service import BuildError
+from compose.core.service import ConvergenceStrategy
+from compose.core.service import ImageType
+from compose.core.service import NeedsBuildError
 
 if not IS_WINDOWS_PLATFORM:
     from dockerpty.pty import PseudoTerminal, RunOperation, ExecOperation
