+++ tests/unit/cli/main_test.py
@@ -11,7 +11,7 @@ from compose.cli.main import convergence_strategy_from_opts
 from compose.cli.main import filter_containers_to_service_names
 from compose.cli.main import setup_console_handler
 from compose.core import container
-from compose.service import ConvergenceStrategy
+from compose.core.service import ConvergenceStrategy
 from tests import mock
